// src/lib/types/colors.ts
export type ColorName =
	| 'red'
	| 'orange'
	| 'amber'
	| 'yellow'
	| 'lime'
	| 'green'
	| 'emerald'
	| 'teal'
	| 'cyan'
	| 'sky'
	| 'blue'
	| 'indigo'
	| 'violet'
	| 'purple'
	| 'fuchsia'
	| 'pink'
	| 'rose'
	| 'slate';
export type HexColor =
	| '#ef4444'
	| '#f97316'
	| '#f59e0b'
	| '#eab308'
	| '#84cc16'
	| '#22c55e'
	| '#10b981'
	| '#14b8a6'
	| '#06b6d4'
	| '#0ea5e9'
	| '#3b82f6'
	| '#6366f1'
	| '#8b5cf6'
	| '#a855f7'
	| '#d946ef'
	| '#ec4899'
	| '#f43f5e'
	| '#64748b';
export type TwClasses = {
	bg: string;
	text: string;
	border: string;
	ring: string;
	hover: string;
	indicator: string;
};